+++ .gitignore
@@ -2,4 +2,4 @@
 /.idea
 /.vscode
 /vendor
-/image-cleanup
+/seiso
