+++ .goreleaser.yml
@@ -31,8 +31,8 @@ snapshot:
 
 dockers:
 - image_templates:
-  - "docker.io/appuio/image-cleanup:v{{ .Version }}"
-  - "docker.io/appuio/image-cleanup:v{{ .Major }}"
+  - "docker.io/appuio/seiso:v{{ .Version }}"
+  - "docker.io/appuio/seiso:v{{ .Major }}"
 
 changelog:
   sort: asc
@@ -45,4 +45,4 @@ changelog:
 release:
   github:
     owner: appuio
-    name: image-cleanup
+    name: seiso
