+++ Dockerfile
@@ -1,9 +1,9 @@
 FROM docker.io/library/alpine:3.11 as runtime
 
-ENTRYPOINT ["image-cleanup"]
+ENTRYPOINT ["seiso"]
 
 RUN \
     apk add --no-cache curl bash
 
-COPY image-cleanup /usr/bin/image-cleanup
+COPY seiso /usr/bin/
 USER 1000:0
