+++ cmd/history.go
@@ -2,10 +2,10 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/appuio/image-cleanup/cfg"
-	"github.com/appuio/image-cleanup/pkg/cleanup"
-	"github.com/appuio/image-cleanup/pkg/git"
-	"github.com/appuio/image-cleanup/pkg/openshift"
+	"github.com/appuio/seiso/cfg"
+	"github.com/appuio/seiso/pkg/cleanup"
+	"github.com/appuio/seiso/pkg/git"
+	"github.com/appuio/seiso/pkg/openshift"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -27,7 +27,7 @@ var (
 )
 
 func init() {
-	rootCmd.AddCommand(historyCmd)
+	imagesCmd.AddCommand(historyCmd)
 	defaults := cfg.NewDefaultConfig()
 
 	addCommonFlagsForGit(historyCmd, defaults)
